+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -40,6 +40,9 @@ from utils.column_names.ind_cqc_pipeline_columns import (
 from utils.column_names.ind_cqc_pipeline_columns import (
     NullGroupedProviderColumns as NGPcol,
 )
+from utils.column_names.ind_cqc_pipeline_columns import (
+    PrimaryServiceRateOfChangeColumns as ROC_TempCol,
+)
 from utils.column_values.categorical_column_values import JobGroupLabels
 from utils.value_labels.ascwds_worker.ascwds_worker_jobgroup_dictionary import (
     AscwdsWorkerValueLabelsJobGroup,
@@ -545,6 +548,7 @@ class ValidateImputedIndCqcAscwdsAndPir:
             (IndCQC.ascwds_filled_posts_source, pl.String()),
             (IndCQC.ascwds_filled_posts_dedup_clean, pl.Float64()),
             (IndCQC.pir_people_directly_employed_dedup, pl.Int64()),
+            (IndCQC.ascwds_rate_of_change_trendline_model, pl.Float32()),
             (IndCQC.pir_people_directly_employed_cleaned, pl.Int64()),
             (IndCQC.filled_posts_per_bed_ratio, pl.Float64()),
         ]
@@ -1636,6 +1640,60 @@ class ModelExtrapolation:
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.ascwds_pir_merged: pl.Float32,
         ExtrapCol.previous_value: pl.Float32,
+    }
+
+
+@dataclass
+class ModelRateOfChangeSchemas:
+    expected_model_roc_trendline_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.care_home: pl.String,
+        IndCQC.number_of_beds: pl.Int32,
+        IndCQC.primary_service_type: pl.String,
+        IndCQC.combined_ratio_and_filled_posts: pl.Float32,
+        IndCQC.care_home_status_count: pl.Int32,
+        IndCQC.ascwds_rate_of_change_trendline_model: pl.Float32,
+    }
+
+    calculate_rolling_sums_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.primary_service_type: pl.String,
+        IndCQC.number_of_beds_banded_roc: pl.Int32,
+        IndCQC.cqc_location_import_date: pl.Date,
+        ROC_TempCol.current_period_cleaned: pl.Float32,
+        ROC_TempCol.previous_period_cleaned: pl.Float32,
+    }
+    expected_calculate_rolling_sums_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.primary_service_type: pl.String,
+        IndCQC.number_of_beds_banded_roc: pl.Int32,
+        IndCQC.cqc_location_import_date: pl.Date,
+        ROC_TempCol.rolling_current_sum: pl.Float32,
+        ROC_TempCol.rolling_previous_sum: pl.Float32,
+    }
+
+    expected_clean_non_residential_rate_of_change_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.care_home: pl.String,
+        IndCQC.cqc_location_import_date: pl.Date,
+        ROC_TempCol.previous_period_interpolated: pl.Float32,
+        ROC_TempCol.current_period_interpolated: pl.Float32,
+        ROC_TempCol.previous_period_cleaned: pl.Float32,
+        ROC_TempCol.current_period_cleaned: pl.Float32,
+    }
+
+    calculate_trendline_schema = {
+        IndCQC.primary_service_type: pl.String,
+        IndCQC.number_of_beds_banded_roc: pl.Int32,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.single_period_rate_of_change: pl.Float64,
+    }
+    expected_calculate_trendline_schema = {
+        IndCQC.primary_service_type: pl.String,
+        IndCQC.number_of_beds_banded_roc: pl.Int32,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.ascwds_rate_of_change_trendline_model: pl.Float64,
     }